create dynamic details page at pages/my-trips/[slug.js]
