+++ components/TripDay/index.js
@@ -1,8 +1,8 @@
 export default function TripDay({ title, activities }) {
   return (
-    <>
+    <li>
       <h3>{title}</h3>
       <p>{activities}</p>
-    </>
+    </li>
   );
 }
